100 Days of SwiftUI - Day 59
Project 12 - Part three

Add a customizable sort order option: by name or by amount.
